Rename init.c to utils.c and Add check_range, check_rgb functions
