+++ Makefile
@@ -6,9 +6,8 @@ RM = rm -f
 
 SRCS = minirt.c \
 	   mlx.c \
-	   init.c \
+	   utils.c \
 	   parse/parse.c \
-	   parse/check_arg.c \
 	   parse/gen_env.c \
 	   parse/gen_objs.c \
 	   get_next_line/get_next_line.c \
